Separate out the program loading stuff into a PROGRAM_LOADER so I can implement extra loaders.
